refactor: share wait frame calculation
